requirements.txt
tensorflow~=2.20.0
numpy>=1.26.0,<2.3.0
flask==3.0.0
flask-cors==4.0.0
pillow>=10.4.0
matplotlib
scikit-learn
